type corrections

+++ pyrealm/core/utilities.py
@@ -24,7 +24,7 @@ but...
   >>> import numpy as np
   >>> x = np.ma.masked_array([1,2,np.nan, 4], mask=[1,0,0,1])
   >>> x.mean()
-  nan
+  np.float64(nan)
   >>> x = np.ma.masked_array([1,2,np.nan, 4], mask=[1,0,1,0])
   >>> x.mean()
   3.0
@@ -282,7 +282,7 @@ def bounds_mask(
     Examples:
         >>> vals = np.array([-15, 20, 30, 124], dtype=float)
         >>> np.nansum(vals)
-        159.0
+        np.float64(159.0)
         >>> vals_c = bounds_mask(vals, 0, 100, label='temperature')
         >>> np.nansum(vals_c)
         50.0
